updated load and unload script

+++ unload.sh
@@ -3,6 +3,8 @@ set -x
 
 BUILD_NAME=bpf_sockops
 
-sudo bpftool cgroup detach /sys/fs/cgroup/ sock_ops pinned /sys/fs/bpf/${BUILD_NAME}
-sudo unlink /sys/fs/bpf/${BUILD_NAME}
-sudo rm -rf /sys/fs/bpf/*
+sudo bpftool cgroup detach /sys/fs/cgroup/unified/${BUILD_NAME} sock_ops pinned /sys/fs/bpf/${BUILD_NAME}
+
+if [ -e "/sys/fs/bpf/bpf_sockops" ]; then
+	sudo rm -rf /sys/fs/bpf/*
+fi
